requirements.txt
Flask-JWT-Extended
flask-smorest
Flask-SQLAlchemy
python-dotenv
marshmallow
gunicorn
passlib